requirements.txt
# 核心依赖
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# 测试依赖
pytest>=7.0.0
aiohttp>=3.9.0

# 可视化依赖（可选）
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0

# 高级加密（可选）
# pycryptodome>=3.19.0

# 生产环境依赖
requests>=2.31.0
redis>=5.0.0
python-dotenv>=1.0.0
psutil>=5.9.0

# 监控和日志（可选）
# sentry-sdk>=1.40.0

# 飞书开放平台SDK（用于真实API调用）
lark-oapi>=1.0.0